refactor(notification): 🛠️ move CreateNotification and SendNotification message classes with unit tests
- Introduced CreateNotification, NotificationCreated, and NotificationCreationFailed message classes to handle notification creation events.
- Added SendNotification, NotificationSent, and NotificationSendFailed classes for managing notification sending events.
- Implemented comprehensive unit tests for both CreateNotification and SendNotification functionalities, ensuring correct payload and metadata handling.
- Enhanced type safety and structure in the notification messaging system, promoting better organization and maintainability.
