+++ packages/notification/src/lib/notification/domain/messages/create-notification.ts
@@ -1,9 +1,9 @@
 import { Message } from '@mostval/common';
 import { TNotificationMetadata, TNotificationMetadataWhenSendFailed } from './send-notification';
-import { Notification, TNotificationProps } from '../../domain';
-import { NotificationError } from '../../domain/error';
+import { Notification, TNotificationProps } from '..';
+import { NotificationError } from '../error';
 
-export class CreateNotification<T> extends Message {
+export class CreateNotification<T> extends Message<TNotificationProps<T>> {
     constructor(payload: TNotificationProps<T>, metadata: TNotificationMetadata) {
         super(payload, metadata);
     }
@@ -12,13 +12,13 @@ export class CreateNotification<T> extends Message {
 export class CreateNotificationError extends NotificationError {
 }
 
-export class NotificationCreated<T> extends Message {
+export class NotificationCreated<T> extends Message<Notification<T>> {
     constructor(payload: Notification<T>, metadata: TNotificationMetadata) {
         super(payload, metadata);
     }
 }
 
-export class NotificationCreationFailed<T> extends Message {
+export class NotificationCreationFailed<T> extends Message<TNotificationProps<T>> {
     constructor(payload: TNotificationProps<T>, metadata: TNotificationMetadataWhenSendFailed<T>) {
         super(payload, metadata);
     }
