+++ packages/notification/src/lib/notification/domain/messages/send-notification.ts
@@ -1,6 +1,6 @@
 import { IMetadata, Message } from '@mostval/common';
-import { Notification } from '../../domain';
-import { NotificationError } from '../../domain/error';
+import { Notification } from '..';
+import { NotificationError } from '../error';
 
 export type TNotificationMetadata = IMetadata
 
@@ -10,19 +10,19 @@ export type TNotificationMetadataWhenSendFailed<T> = TNotificationMetadataWhenSe
     readonly error: NotificationError;
 }
 
-export class SendNotification<T> extends Message {
+export class SendNotification<T> extends Message<Notification<T>> {
     constructor(payload: Notification<T>, metadata: TNotificationMetadataWhenSending<T>) {
         super(payload, metadata);
     }
 }
 
-export class NotificationSent<T> extends Message {
+export class NotificationSent<T> extends Message<Notification<T>> {
     constructor(payload: Notification<T>, metadata: TNotificationMetadata) {
         super(payload, metadata);
     }
 }
 
-export class NotificationSendFailed<T> extends Message {
+export class NotificationSendFailed<T> extends Message<Notification<T>> {
     constructor(payload: Notification<T>, metadata: TNotificationMetadataWhenSendFailed<T>) {
         super(payload, metadata);
     }
